requirements.txt
telethon>=1.34.0
requests>=2.31.0
python-dotenv>=1.0.0
pyTelegramBotAPI>=4.14.0
# Flask и pytz устанавливаются отдельно для flask/ приложения
# см. flask/requirements.txt
